conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
abseil-cpp=20210324.2=h9c3ff4c_0
absl-py=1.0.0=pyhd8ed1ab_0
addict=2.4.0=pyhd8ed1ab_2
aiohttp=3.7.4.post0=py38h497a2fe_1
arrow-cpp=5.0.0=py38hd2b13db_8_cuda
arrow-cpp-proc=3.0.0=cuda
async-timeout=3.0.1=py_1000
attrs=21.4.0=pyhd8ed1ab_0
aws-c-cal=0.5.11=h95a6274_0
aws-c-common=0.6.2=h7f98852_0
aws-c-event-stream=0.2.7=h3541f99_13
aws-c-io=0.10.5=hfb6a706_0
aws-checksums=0.1.11=ha31a3da_7
aws-sdk-cpp=1.8.186=hb4091e7_3
blas=1.0=mkl
blinker=1.4=py_1
bokeh=2.4.2=py38h578d9bd_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h497a2fe_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.14.6=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_2
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.0.4=py38h578d9bd_0
cloudpickle=2.0.0=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.2=py38h2b5fc30_0
cuda-python=11.6.1=py38h3fd9d12_0
cudatoolkit=11.5.1=hcf5317a_9
cudf=22.02.00=cuda_11_py38_g774d859fef_0
cudnn=8.2.1.32=h86fa8c9_0
cuml=22.02.00=cuda11_py38_g679e4fa21_0
cupy=10.2.0=py38h405e1b6_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.2=py38h497a2fe_1
dask=2022.1.0=pyhd8ed1ab_0
dask-core=2022.1.0=pyhd8ed1ab_0
dask-cudf=22.02.00=cuda_11_py38_g774d859fef_0
dbus=1.13.18=hb2f20db_0
distributed=2022.1.0=py38h578d9bd_0
dlpack=0.5=h9c3ff4c_0
expat=2.4.4=h295c915_0
faiss=1.7.0=py38cuda112h0532a0c_2_cuda
faiss-gpu=1.7.0=h500d9d2_8
faiss-proc=1.0.0=cuda
fastavro=1.4.10=py38h0a891b7_0
fastrlock=0.8=py38h709712a_1
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fsspec=2022.2.0=pyhd8ed1ab_0
future=0.18.2=py38h578d9bd_4
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
glog=0.5.0=h48cff8f_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyh6c4a22f_1
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpc-cpp=1.40.0=h05f19cf_1
grpcio=1.38.1=py38hdd6454d_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
heapdict=1.0.1=py_0
icu=58.2=he6710b0_3
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.3=py38h578d9bd_0
intel-openmp=2021.4.0=h06a4308_3561
jbig=2.1=h7f98852_2003
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.0=py38h43d8883_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=2.2.1=h9c3ff4c_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h7f98852_6
libbrotlidec=1.0.9=h7f98852_6
libbrotlienc=1.0.9=h7f98852_6
libcblas=3.9.0=12_linux64_mkl
libcudf=22.02.00=cuda11_g774d859fef_0
libcuml=22.02.00=cuda11_g679e4fa21_0
libcumlprims=22.02.00=cuda11_g2207299_0
libcurl=7.79.1=h2574ce0_1
libcusolver=11.3.3.112=hc34d849_0
libdeflate=1.7=h7f98852_5
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libfaiss=1.7.0=cuda112h5bea7ad_8_cuda
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_14
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libhwloc=2.3.0=h5e5b7d1_1
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
libllvm11=11.1.0=h3826bc1_1
libnghttp2=1.43.0=h812cca2_1
libpng=1.6.37=hbc83047_0
libprotobuf=3.18.0=h780b84a_1
librmm=22.02.00=cuda11_ge3e3215_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=he4da1e4_14
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=he6d91bd_0
libtiff=4.3.0=hf544144_1
libunistring=0.9.10=h27cfd23_0
libutf8proc=2.7.0=h7f98852_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvm-openmp=12.0.1=h4bd325d_1
llvmlite=0.37.0=py38h4630a5e_0
locket=0.2.0=py_2
lz4-c=1.9.3=h295c915_1
magma=2.5.4=h6103c52_2
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.1.1=py38h0a891b7_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h5c7f4ab_1
mkl=2021.4.0=h06a4308_640
msgpack-python=1.0.3=py38h1fd1430_0
multidict=6.0.2=py38h497a2fe_0
munkres=1.1.4=py_0
nccl=2.12.7.1=h0800d71_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=h4bd325d_1
numba=0.54.1=py38h4bf6c61_0
numpy=1.20.3=py38h9894fe3_1
nvtx=0.2.3=py38h497a2fe_1
oauthlib=3.2.0=pyhd8ed1ab_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1l=h7f98852_0
orc=1.7.0=h68e2c4e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py38h43a58ef_0
parquet-cpp=1.5.1=2
partd=1.2.0=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pillow=8.3.2=py38h8e6f84c_0
pip=22.0.4=pyhd8ed1ab_0
protobuf=3.18.0=py38h709712a_0
psutil=5.9.0=py38h497a2fe_0
ptxcompiler=0.2.0=py38h98f4b32_0
pyarrow=5.0.0=py38hed47224_8_cuda
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py38h578d9bd_4
python=3.8.10=h49503c6_1_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-lmdb=1.3.0=py38h709712a_0
python_abi=3.8=2_cp38
pytorch=1.9.1=cuda112py38h4f2a933_3
pytorch-gpu=1.9.1=cuda112py38h0bbbad9_3
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py38h497a2fe_3
qt=5.9.7=h5867ecd_1
re2=2021.09.01=h9c3ff4c_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rmm=22.02.00=cuda11_py38_ge3e3215_0_has_cma
rsa=4.8=pyhd8ed1ab_0
s2n=1.0.10=h9b69904_0
scikit-learn=1.0.2=py38h1561384_0
scipy=1.5.3=py38h828c644_0
setuptools=59.8.0=py38h578d9bd_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h9b69904_2
snappy=1.1.8=he1b5a44_3
sortedcontainers=2.4.0=pyhd8ed1ab_0
spdlog=1.8.5=h4bd325d_1
sqlite=3.37.2=hc218d9a_0
tbb=2021.5.0=hd09550d_0
tblib=1.7.0=pyhd8ed1ab_0
tensorboard=2.8.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.0=py38h3e25421_1
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
toolz=0.11.2=pyhd8ed1ab_0
torchvision=0.10.1=py38cuda112h04b465a_0_cuda
tornado=6.1=py38h497a2fe_2
tqdm=4.63.0=pyhd8ed1ab_0
treelite=2.2.1=py38hdd725b4_2
treelite-runtime=2.2.1=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
ucx=1.12.0+gd367332=cuda11.2_0
ucx-proc=1.0.0=gpu
ucx-py=0.24.0=py38_gd367332_0
urllib3=1.26.8=pyhd8ed1ab_1
werkzeug=2.0.3=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py38h497a2fe_1
zict=2.1.0=pyhd8ed1ab_0
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h7f8727e_4
zstd=1.5.0=ha95c52a_0
